[Quartermaster] refactor: Split StatsPanel.axaml.cs into partial classes (#1126)
Split 830-LOC file into 4 focused partial classes:
- StatsPanel.axaml.cs (314 LOC) - Core initialization/loading
- StatsPanel.Abilities.cs (245 LOC) - Ability score handling
- StatsPanel.Combat.cs (209 LOC) - HP/AC/BAB/CR
- StatsPanel.Saves.cs (101 LOC) - Saving throws

All 87 tests pass. No file exceeds 500 LOC.
