+++ Quartermaster/Quartermaster/Views/Panels/StatsPanel.axaml.cs
@@ -9,6 +9,9 @@ using Radoub.Formats.Utc;
 
 namespace Quartermaster.Views.Panels;
 
+/// <summary>
+/// StatsPanel - Core initialization, loading, and field definitions.
+/// </summary>
 public partial class StatsPanel : UserControl
 {
     private CreatureDisplayService? _displayService;
@@ -258,525 +261,6 @@ public partial class StatsPanel : UserControl
 
         // Defer clearing _isLoading until after dispatcher processes queued ValueChanged events
         Avalonia.Threading.Dispatcher.UIThread.Post(() => _isLoading = false, Avalonia.Threading.DispatcherPriority.Background);
-    }
-
-    private void OnCRAdjustValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-
-        _currentCreature.CRAdjust = (int)(e.NewValue ?? 0);
-        UpdateCrTotalDisplay();
-        CRAdjustChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    /// <summary>
-    /// Updates the total CR display (Calculated + Adjustment).
-    /// </summary>
-    private void UpdateCrTotalDisplay()
-    {
-        if (_crTotalDisplay == null || _currentCreature == null) return;
-
-        float totalCr = _currentCreature.ChallengeRating + _currentCreature.CRAdjust;
-        SetText(_crTotalDisplay, totalCr.ToString("0.##"));
-    }
-
-    private void OnNaturalAcValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-
-        _currentCreature.NaturalAC = (byte)(e.NewValue ?? 0);
-        RecalculateTotalAc();
-        NaturalAcChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnBaseHpValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-
-        short newBaseHp = (short)(e.NewValue ?? 1);
-        _currentCreature.HitPoints = newBaseHp;
-
-        // Recalculate MaxHP = BaseHP + Con contribution
-        RecalculateMaxHp();
-
-        HitPointsChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void RecalculateMaxHp()
-    {
-        if (_currentCreature == null) return;
-
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int conTotal = _currentCreature.Con + racialMods.Con;
-        int conBonus = CreatureDisplayService.CalculateAbilityBonus(conTotal);
-        int totalLevel = _currentCreature.ClassList.Sum(c => c.ClassLevel);
-        int conHpContribution = conBonus * totalLevel;
-
-        // Calculate new MaxHP
-        int newMaxHp = _currentCreature.HitPoints + conHpContribution;
-        _currentCreature.MaxHitPoints = (short)Math.Max(1, newMaxHp);
-
-        // Set CurrentHP = MaxHP (creatures spawn at full health)
-        _currentCreature.CurrentHitPoints = _currentCreature.MaxHitPoints;
-
-        // Update displays
-        SetText(_maxHpValue, _currentCreature.MaxHitPoints.ToString());
-        SetText(_conHpBonus, CreatureDisplayService.FormatBonus(conHpContribution));
-    }
-
-    private void WireAbilityScoreEvents()
-    {
-        if (_strBase != null) _strBase.ValueChanged += OnStrValueChanged;
-        if (_dexBase != null) _dexBase.ValueChanged += OnDexValueChanged;
-        if (_conBase != null) _conBase.ValueChanged += OnConValueChanged;
-        if (_intBase != null) _intBase.ValueChanged += OnIntValueChanged;
-        if (_wisBase != null) _wisBase.ValueChanged += OnWisValueChanged;
-        if (_chaBase != null) _chaBase.ValueChanged += OnChaValueChanged;
-    }
-
-    private void OnStrValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Str = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Str", _currentCreature.Str);
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnDexValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Dex = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Dex", _currentCreature.Dex);
-        UpdateDexAcDisplay(); // Dex affects AC
-        UpdateSavingThrows(); // Dex affects Reflex save
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnConValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Con = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Con", _currentCreature.Con);
-        UpdateHitPointsDisplay(); // Con affects HP
-        UpdateSavingThrows(); // Con affects Fortitude save
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnIntValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Int = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Int", _currentCreature.Int);
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnWisValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Wis = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Wis", _currentCreature.Wis);
-        UpdateSavingThrows(); // Wis affects Will save
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnChaValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.Cha = (byte)(e.NewValue ?? 10);
-        UpdateAbilityDisplay("Cha", _currentCreature.Cha);
-        UpdateAbilityPointsSummary();
-        AbilityScoresChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    /// <summary>
-    /// Refreshes the ability points summary display.
-    /// Call this when class levels change externally.
-    /// </summary>
-    public void RefreshAbilityPointsSummary()
-    {
-        UpdateAbilityPointsSummary();
-    }
-
-    /// <summary>
-    /// Updates the ability points summary display.
-    /// Shows expected vs used ability points from level-up bonuses.
-    /// Characters gain 1 ability point at levels 4, 8, 12, 16, 20, 24, 28, 32, 36, 40.
-    /// </summary>
-    private void UpdateAbilityPointsSummary()
-    {
-        if (_abilityPointsSummary == null || _currentCreature == null)
-        {
-            SetText(_abilityPointsSummary, "");
-            return;
-        }
-
-        // Calculate total level
-        int totalLevel = _currentCreature.ClassList.Sum(c => c.ClassLevel);
-
-        // Expected ability points from leveling: 1 point every 4 levels
-        int expectedPoints = totalLevel / 4;
-
-        // Calculate used points by comparing current base scores against starting values
-        // NWN point-buy starts at 8 for all abilities, and players distribute 30 points
-        // For creatures, we estimate starting scores as 8 (minimum typical point-buy value)
-        // Used points = sum of (current - 8) for all abilities, clamped to non-negative
-        var abilityPointsInfo = CalculateAbilityPointsFromLeveling();
-        int usedPoints = abilityPointsInfo.UsedPoints;
-
-        // Only show for creatures with levels (not level 0 templates)
-        if (totalLevel == 0)
-        {
-            SetText(_abilityPointsSummary, "");
-            return;
-        }
-
-        // Format: "Level Points: X/Y" where X = used, Y = expected
-        string text = $"Level Points: {usedPoints}/{expectedPoints}";
-
-        // Change color based on status
-        if (_abilityPointsSummary != null)
-        {
-            if (usedPoints > expectedPoints)
-            {
-                // Over-allocated (possibly from items, buffs, or manual editing)
-                _abilityPointsSummary.Foreground = this.FindResource("ThemeWarning") as Avalonia.Media.IBrush
-                    ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-                text += " (over)";
-            }
-            else if (usedPoints < expectedPoints)
-            {
-                // Under-allocated (points available)
-                _abilityPointsSummary.Foreground = this.FindResource("ThemeInfo") as Avalonia.Media.IBrush
-                    ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-                text += $" ({expectedPoints - usedPoints} available)";
-            }
-            else
-            {
-                // Exactly matched
-                _abilityPointsSummary.Foreground = this.FindResource("ThemeSuccess") as Avalonia.Media.IBrush
-                    ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-            }
-        }
-
-        SetText(_abilityPointsSummary, text);
-    }
-
-    /// <summary>
-    /// Calculates ability points from leveling by estimating how many points are beyond point-buy max.
-    /// NWN point-buy max per ability is 18 at creation.
-    /// Level-up points add +1 directly, bypassing point-buy cost escalation.
-    /// We count points above 18 as definite level-up points, plus estimate for 17-18 range.
-    /// </summary>
-    private (int UsedPoints, int EstimatedStartingTotal) CalculateAbilityPointsFromLeveling()
-    {
-        if (_currentCreature == null)
-            return (0, 0);
-
-        // Count ability points that must have come from level-ups
-        // Any score above 18 definitely came from level-up (can't point-buy above 18)
-        // For scores 17-18, we estimate based on how common those starting values are
-        int levelUpPoints = 0;
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Str);
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Dex);
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Con);
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Int);
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Wis);
-        levelUpPoints += CountLevelUpPointsForAbility(_currentCreature.Cha);
-
-        // For display purposes, calculate what the starting total would have been
-        int currentTotal = _currentCreature.Str + _currentCreature.Dex + _currentCreature.Con +
-                          _currentCreature.Int + _currentCreature.Wis + _currentCreature.Cha;
-        int estimatedStartingTotal = currentTotal - levelUpPoints;
-
-        return (levelUpPoints, estimatedStartingTotal);
-    }
-
-    /// <summary>
-    /// Estimates how many level-up points contributed to a single ability score.
-    /// Points above 18 are definitely from level-ups (can't point-buy above 18).
-    /// </summary>
-    private static int CountLevelUpPointsForAbility(int score)
-    {
-        // Scores above 18 must have come from level-up points
-        // (Point-buy caps at 18)
-        if (score > 18)
-            return score - 18;
-
-        // Scores 18 and below could be entirely from point-buy
-        return 0;
-    }
-
-    /// <summary>
-    /// Updates the expected HP summary display.
-    /// Shows comparison of base HP against expected range from class hit dice.
-    /// </summary>
-    private void UpdateExpectedHpSummary(UtcFile creature)
-    {
-        if (_expectedHpSummary == null || _displayService == null)
-        {
-            SetText(_expectedHpSummary, "");
-            return;
-        }
-
-        int totalLevel = creature.ClassList.Sum(c => c.ClassLevel);
-        if (totalLevel == 0)
-        {
-            SetText(_expectedHpSummary, "");
-            Avalonia.Controls.ToolTip.SetTip(_expectedHpSummary, null);
-            return;
-        }
-
-        // Calculate expected HP range from dice rolls
-        var (minHp, avgHp, maxHp) = _displayService.CalculateExpectedHpRange(creature);
-        int baseHp = creature.HitPoints;
-
-        // Set tooltip with full explanation
-        string tooltip = $"Expected hit points are between {minHp} and {maxHp} with an average of {avgHp}.\nFirst level gets max die, subsequent levels roll 1 to hitDie.";
-        Avalonia.Controls.ToolTip.SetTip(_expectedHpSummary, tooltip);
-
-        // Format display text
-        string text;
-        if (baseHp < minHp)
-        {
-            // Below minimum - unusual
-            _expectedHpSummary.Foreground = this.FindResource("ThemeWarning") as Avalonia.Media.IBrush
-                ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-            text = $"Expected: {minHp}-{maxHp} (low)";
-        }
-        else if (baseHp > maxHp)
-        {
-            // Above maximum - has bonuses
-            _expectedHpSummary.Foreground = this.FindResource("ThemeInfo") as Avalonia.Media.IBrush
-                ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-            text = $"Expected: {minHp}-{maxHp} (+{baseHp - maxHp})";
-        }
-        else if (baseHp >= avgHp)
-        {
-            // Good rolls
-            _expectedHpSummary.Foreground = this.FindResource("ThemeSuccess") as Avalonia.Media.IBrush
-                ?? this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-            text = $"Expected: {minHp}-{maxHp}";
-        }
-        else
-        {
-            // Below average but valid
-            _expectedHpSummary.Foreground = this.FindResource("SystemControlForegroundBaseMediumBrush") as Avalonia.Media.IBrush;
-            text = $"Expected: {minHp}-{maxHp}";
-        }
-
-        SetText(_expectedHpSummary, text);
-    }
-
-    private void UpdateAbilityDisplay(string ability, byte baseValue)
-    {
-        if (_currentCreature == null) return;
-
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int racialMod = ability switch
-        {
-            "Str" => racialMods.Str,
-            "Dex" => racialMods.Dex,
-            "Con" => racialMods.Con,
-            "Int" => racialMods.Int,
-            "Wis" => racialMods.Wis,
-            "Cha" => racialMods.Cha,
-            _ => 0
-        };
-
-        int total = baseValue + racialMod;
-        int bonus = CreatureDisplayService.CalculateAbilityBonus(total);
-
-        var (racialCtrl, totalCtrl, bonusCtrl) = ability switch
-        {
-            "Str" => (_strRacial, _strTotal, _strBonus),
-            "Dex" => (_dexRacial, _dexTotal, _dexBonus),
-            "Con" => (_conRacial, _conTotal, _conBonus),
-            "Int" => (_intRacial, _intTotal, _intBonus),
-            "Wis" => (_wisRacial, _wisTotal, _wisBonus),
-            "Cha" => (_chaRacial, _chaTotal, _chaBonus),
-            _ => (null, null, null)
-        };
-
-        SetText(racialCtrl, CreatureDisplayService.FormatBonus(racialMod));
-        SetText(totalCtrl, total.ToString());
-        SetText(bonusCtrl, CreatureDisplayService.FormatBonus(bonus));
-    }
-
-    private void UpdateHitPointsDisplay()
-    {
-        // Delegate to the shared recalculation logic
-        RecalculateMaxHp();
-    }
-
-    private void UpdateDexAcDisplay()
-    {
-        if (_currentCreature == null) return;
-
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int dexTotal = _currentCreature.Dex + racialMods.Dex;
-        int dexBonus = CreatureDisplayService.CalculateAbilityBonus(dexTotal);
-
-        SetText(_dexAcValue, CreatureDisplayService.FormatBonus(dexBonus));
-
-        // Dex change affects Total AC
-        RecalculateTotalAc();
-    }
-
-    private void RecalculateTotalAc()
-    {
-        if (_currentCreature == null) return;
-
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int dexTotal = _currentCreature.Dex + racialMods.Dex;
-        int dexBonus = CreatureDisplayService.CalculateAbilityBonus(dexTotal);
-        int sizeAcMod = _displayService?.GetSizeAcModifier(_currentCreature.AppearanceType) ?? 0;
-
-        // Total AC = 10 (base) + Natural AC + Dex Bonus + Size Mod
-        int totalAc = 10 + _currentCreature.NaturalAC + dexBonus + sizeAcMod;
-        SetText(_totalAcValue, totalAc.ToString());
-    }
-
-    private void UpdateSavingThrows()
-    {
-        if (_currentCreature == null) return;
-
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-
-        int conTotal = _currentCreature.Con + racialMods.Con;
-        int dexTotal = _currentCreature.Dex + racialMods.Dex;
-        int wisTotal = _currentCreature.Wis + racialMods.Wis;
-
-        int conBonus = CreatureDisplayService.CalculateAbilityBonus(conTotal);
-        int dexBonus = CreatureDisplayService.CalculateAbilityBonus(dexTotal);
-        int wisBonus = CreatureDisplayService.CalculateAbilityBonus(wisTotal);
-
-        LoadSavingThrow(_fortBase, _fortAbility, _fortTotal, _currentCreature.FortBonus, conBonus);
-        LoadSavingThrow(_refBase, _refAbility, _refTotal, _currentCreature.RefBonus, dexBonus);
-        LoadSavingThrow(_willBase, _willAbility, _willTotal, _currentCreature.WillBonus, wisBonus);
-    }
-
-    /// <summary>
-    /// Sets the equipped items for combat stat calculations.
-    /// Call this after loading creature and populating inventory.
-    /// </summary>
-    public void SetEquippedItems(IEnumerable<UtiFile?> items)
-    {
-        _equippedItems = items;
-        UpdateBabDisplay();
-    }
-
-    private void UpdateBabDisplay()
-    {
-        if (_currentCreature == null || _displayService == null)
-        {
-            SetText(_babValue, "+0");
-            SetText(_babBreakdown, "");
-            return;
-        }
-
-        var combatStats = _displayService.CalculateCombatStats(_currentCreature, _equippedItems);
-
-        // Display total BAB
-        SetText(_babValue, CreatureDisplayService.FormatBonus(combatStats.TotalBab));
-
-        // Display breakdown if equipment bonus present
-        if (combatStats.EquipmentBonus > 0)
-        {
-            SetText(_babBreakdown, $"({combatStats.BaseBab} base + {combatStats.EquipmentBonus} equip)");
-        }
-        else
-        {
-            SetText(_babBreakdown, $"(from class levels)");
-        }
-    }
-
-    private void LoadAbilityScore(NumericUpDown? baseCtrl, TextBlock? racialCtrl, TextBlock? totalCtrl, TextBlock? bonusCtrl,
-        byte baseValue, int racialMod)
-    {
-        int total = baseValue + racialMod;
-        int bonus = CreatureDisplayService.CalculateAbilityBonus(total);
-
-        if (baseCtrl != null) baseCtrl.Value = baseValue;
-        SetText(racialCtrl, CreatureDisplayService.FormatBonus(racialMod));
-        SetText(totalCtrl, total.ToString());
-        SetText(bonusCtrl, CreatureDisplayService.FormatBonus(bonus));
-    }
-
-    private void LoadSavingThrow(NumericUpDown? baseCtrl, TextBlock? abilityCtrl, TextBlock? totalCtrl,
-        short baseValue, int abilityBonus)
-    {
-        int total = baseValue + abilityBonus;
-
-        if (baseCtrl != null) baseCtrl.Value = baseValue;
-        SetText(abilityCtrl, CreatureDisplayService.FormatBonus(abilityBonus));
-        SetText(totalCtrl, CreatureDisplayService.FormatBonus(total));
-    }
-
-    private void WireSavingThrowEvents()
-    {
-        if (_fortBase != null) _fortBase.ValueChanged += OnFortValueChanged;
-        if (_refBase != null) _refBase.ValueChanged += OnRefValueChanged;
-        if (_willBase != null) _willBase.ValueChanged += OnWillValueChanged;
-    }
-
-    private void OnFortValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.FortBonus = (short)(e.NewValue ?? 0);
-        UpdateSavingThrowTotal(_fortBase, _fortAbility, _fortTotal, _currentCreature.FortBonus, GetConBonus());
-        SavingThrowsChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnRefValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.RefBonus = (short)(e.NewValue ?? 0);
-        UpdateSavingThrowTotal(_refBase, _refAbility, _refTotal, _currentCreature.RefBonus, GetDexBonus());
-        SavingThrowsChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void OnWillValueChanged(object? sender, NumericUpDownValueChangedEventArgs e)
-    {
-        if (_isLoading || _currentCreature == null) return;
-        _currentCreature.WillBonus = (short)(e.NewValue ?? 0);
-        UpdateSavingThrowTotal(_willBase, _willAbility, _willTotal, _currentCreature.WillBonus, GetWisBonus());
-        SavingThrowsChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    private void UpdateSavingThrowTotal(NumericUpDown? baseCtrl, TextBlock? abilityCtrl, TextBlock? totalCtrl,
-        short baseValue, int abilityBonus)
-    {
-        int total = baseValue + abilityBonus;
-        SetText(totalCtrl, CreatureDisplayService.FormatBonus(total));
-    }
-
-    private int GetConBonus()
-    {
-        if (_currentCreature == null) return 0;
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int conTotal = _currentCreature.Con + racialMods.Con;
-        return CreatureDisplayService.CalculateAbilityBonus(conTotal);
-    }
-
-    private int GetDexBonus()
-    {
-        if (_currentCreature == null) return 0;
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int dexTotal = _currentCreature.Dex + racialMods.Dex;
-        return CreatureDisplayService.CalculateAbilityBonus(dexTotal);
-    }
-
-    private int GetWisBonus()
-    {
-        if (_currentCreature == null) return 0;
-        var racialMods = _displayService?.GetRacialModifiers(_currentCreature.Race) ?? new RacialModifiers();
-        int wisTotal = _currentCreature.Wis + racialMods.Wis;
-        return CreatureDisplayService.CalculateAbilityBonus(wisTotal);
     }
 
     public void ClearStats()
